Moved boxplot series to subfolder.
